requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiortc==1.6.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.8
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
websockets==12.0
python-dotenv==1.0.0
# Audio processing dependencies
pyaudio==0.2.11
numpy==1.24.3
scipy==1.11.1